Let player see spawn point by pressing "change"
The idea behind this is that when the spawn point is near one of the
edges of the map, there was previously no way to see exactly where you
would spawn. This would get problematic if there were spike bombs etc in
the area. It's simply not fun to spawn and then immediately lose a lot
of health with no way of preventing it.

The spawn point is not displayed automatically because when I discussed
this potential change with players they were concerned that the other
player could see where they would spawn. Having this on a button
mitigates that concern. As a bonus, I don't have to figure out a way of
drawing the worm that makes it clear you haven't spawned yet.